requirements.txt
streamlit
PyPDF2
faiss-cpu
langchain-openai
langchain-community
python-dotenv
